src/recommender-service/requirements.txt
flask==1.1.2
firebase-admin==4.5.3
scikit-learn==0.24.1
gql==2.0.0
numpy==1.20.1
pytest-flask==1.2.0